Refactors the Parameter of jobs.
Previously parameters were stateful instances which could not safely be shared or reused.

We now split the builder from the immutable result "Parameter" which can be made public
visible without the danger of a consumer or job which modifies the global instance
accidentally.
